updat input field
